Moving setLayout functions to utils
